hk_job_fair_aggregator/jobfair-agent/hk_job_fair_aggregator/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.1
pytz==2023.3.post1
python-dateutil==2.8.2
lxml==4.9.3
tenacity==8.2.3
opencc-python-reimplemented==0.1.7
